moved to entity package
