+++ src/main/java/com/diyo/flightbooking/entity/Payment.java
@@ -1,10 +1,13 @@
-package com.diyo.flightbooking;
+package com.diyo.flightbooking.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Entity
 @Getter
